Finished linting step and added raw loops.
Raw loops are integrated into the parser and lexer correctly, but there
needs to be some work on the tests for them. They're marked TODO.

Once this is done, I need to add the tidy workflow.

A PR can then be opened to check cross platform portability, especially
since I touched a ton of casts on the C++ side. Shouldnt be too hard,
and I'll just squash as to not mess with the commit log.

+++ src/util/alphanum.c
@@ -118,7 +118,7 @@ NODISCARD Status strntochr(const char* str, size_t n, uint8_t* out) {
         *out = '\0';
         break;
     default:
-        *out = (uint8_t)escaped;
+        *out = escaped;
         break;
     }
